Tests/LibWeb/Text/expected/css/FontFace-load-urls.txt

hashSans.family: Hash Sans
hashSans.status: unloaded
hashSans.status: loaded
font load failed because: NetworkError: Failed to load font
notExistFont.status: error
